Refactor GameView into smaller SwiftUI components
