+++ UI/GameView.swift
@@ -409,38 +409,6 @@ struct GameView: View {
         .zIndex(2)
     }
 
-    /// 盤面上部に表示する統計テキストの共通レイアウト
-    /// - Parameters:
-    ///   - title: メトリクスのラベル（例: 移動）
-    ///   - value: 表示する数値文字列
-    ///   - accessibilityLabel: VoiceOver に読み上げさせる日本語ラベル
-    ///   - accessibilityValue: VoiceOver 用の値（単位を含めた文章）
-    /// - Returns: モノクロ配色の小型バッジビュー
-    private func statisticBadge(
-        title: String,
-        value: String,
-        accessibilityLabel: String,
-        accessibilityValue: String
-    ) -> some View {
-        VStack(alignment: .leading, spacing: 2) {
-            // 補助ラベルは控えめなトーンで表示
-            Text(title)
-                .font(.caption2)
-                .fontWeight(.medium)
-                // テーマの補助文字色でライト/ダーク双方のコントラストを調整
-                .foregroundColor(theme.statisticTitleText)
-
-            // 主数値は視認性を高めるためサイズとコントラストを強調
-            Text(value)
-                .font(.headline)
-                .foregroundColor(theme.statisticValueText)
-        }
-        // VoiceOver ではカスタムラベルと値を読み上げさせる
-        .accessibilityElement(children: .ignore)
-        .accessibilityLabel(accessibilityLabel)
-        .accessibilityValue(accessibilityValue)
-    }
-
     /// レイアウトに関する最新の実測値をログに残すための不可視ビューを生成
     /// - Parameter context: GeometryReader から抽出したレイアウト情報コンテキスト
     /// - Returns: 画面上には表示されない監視用ビュー
@@ -535,767 +503,12 @@ private extension GameView {
     }
 }
 
-// MARK: - 盤面上部のコントロールバー
-/// 統計バッジと手動操作ボタンをまとめて表示するための補助ビュー
-/// - Note: GameView 本体から UI 細部の責務を切り出し、可読性と再利用性を高める。
-fileprivate struct GameBoardControlRowView: View {
-    /// 共通のテーマ配色
-    let theme: AppTheme
-    /// ゲーム進行とサービス連携を管理する ViewModel
-    @ObservedObject var viewModel: GameViewModel
-
-    var body: some View {
-        ViewThatFits(in: .horizontal) {
-            singleLineLayout
-            stackedLayout
-        }
-        .padding(.horizontal, 16)
-        // PreferenceKey へ高さを伝搬し、GeometryReader 側のレイアウト計算へ反映する
-        .overlay(alignment: .topLeading) {
-            HeightPreferenceReporter<StatisticsHeightPreferenceKey>()
-        }
-    }
-
-    /// 横幅に余裕がある場合に利用する 1 行構成
-    private var singleLineLayout: some View {
-        HStack(alignment: .center, spacing: 12) {
-            flexibleStatisticsContainer()
-            controlButtonCluster()
-        }
-    }
-
-    /// 横幅不足時に統計とボタンを 2 行へ分割するレイアウト
-    private var stackedLayout: some View {
-        VStack(alignment: .leading, spacing: 12) {
-            flexibleStatisticsContainer()
-
-            controlButtonCluster()
-                .frame(maxWidth: .infinity, alignment: .trailing)
-        }
-    }
-
-    /// 統計バッジ群を横幅に合わせてスクロールへフォールバック可能にする
-    private func flexibleStatisticsContainer() -> some View {
-        ViewThatFits(in: .horizontal) {
-            statisticsBadgeContainer()
-            ScrollView(.horizontal, showsIndicators: false) {
-                statisticsBadgeContainer()
-            }
-        }
-        .frame(maxWidth: .infinity, alignment: .leading)
-        .layoutPriority(1)
-    }
-
-    /// 統計バッジを 2 つのグループに分けて表示
-    private func statisticsBadgeContainer() -> some View {
-        HStack(spacing: 12) {
-            scoreStatisticsGroup()
-            supplementaryStatisticsGroup()
-        }
-    }
-
-    /// スコアへ直接影響する指標群
-    private func scoreStatisticsGroup() -> some View {
-        statisticsBadgeGroup {
-            statisticBadge(
-                title: "移動",
-                value: "\(viewModel.moveCount)",
-                accessibilityLabel: "移動回数",
-                accessibilityValue: "\(viewModel.moveCount)回"
-            )
-
-            statisticBadge(
-                title: "ペナルティ",
-                value: "\(viewModel.penaltyCount)",
-                accessibilityLabel: "ペナルティ回数",
-                accessibilityValue: "\(viewModel.penaltyCount)手"
-            )
-
-            statisticBadge(
-                title: "経過時間",
-                value: formattedElapsedTime(viewModel.displayedElapsedSeconds),
-                accessibilityLabel: "経過時間",
-                accessibilityValue: accessibilityElapsedTimeDescription(viewModel.displayedElapsedSeconds)
-            )
-
-            statisticBadge(
-                title: "総合スコア",
-                value: "\(viewModel.displayedScore)",
-                accessibilityLabel: "総合スコア",
-                accessibilityValue: accessibilityScoreDescription(viewModel.displayedScore)
-            )
-        }
-    }
-
-    /// 進行度合いを補足する指標群（残りマスなど）
-    private func supplementaryStatisticsGroup() -> some View {
-        statisticsBadgeGroup {
-            statisticBadge(
-                title: "残りマス",
-                value: "\(viewModel.remainingTiles)",
-                accessibilityLabel: "残りマス数",
-                accessibilityValue: "残り\(viewModel.remainingTiles)マス"
-            )
-        }
-    }
-
-    /// 共通の装飾を適用した統計バッジコンテナ
-    private func statisticsBadgeGroup<Content: View>(@ViewBuilder content: () -> Content) -> some View {
-        HStack(spacing: 12) {
-            content()
-        }
-        .padding(.horizontal, 12)
-        .padding(.vertical, 10)
-        .background(
-            RoundedRectangle(cornerRadius: 12)
-                .fill(theme.statisticBadgeBackground)
-        )
-        .overlay(
-            RoundedRectangle(cornerRadius: 12)
-                .stroke(theme.statisticBadgeBorder, lineWidth: 1)
-        )
-        .accessibilityElement(children: .contain)
-    }
-
-    /// ボタン 3 種を横並びで表示
-    private func controlButtonCluster() -> some View {
-        HStack(spacing: 12) {
-            manualDiscardButton
-            manualPenaltyButton
-            pauseButton
-        }
-    }
-
-    /// 捨て札モード切替ボタン
-    private var manualDiscardButton: some View {
-        let isSelecting = viewModel.isAwaitingManualDiscardSelection
-        let isDisabled = !viewModel.isManualDiscardButtonEnabled && !isSelecting
-
-        return Button {
-            viewModel.toggleManualDiscardSelection()
-        } label: {
-            Image(systemName: "trash")
-                .font(.system(size: 20, weight: .semibold))
-                .foregroundColor(isSelecting ? theme.accentOnPrimary : theme.menuIconForeground)
-                .frame(width: 44, height: 44)
-                .background(
-                    Circle()
-                        .fill(isSelecting ? theme.accentPrimary : theme.menuIconBackground)
-                )
-                .overlay(
-                    Circle()
-                        .stroke(
-                            isSelecting ? theme.accentPrimary.opacity(0.55) : theme.menuIconBorder,
-                            lineWidth: 1
-                        )
-                )
-        }
-        .buttonStyle(.plain)
-        .opacity(isDisabled ? 0.45 : 1.0)
-        .disabled(isDisabled)
-        .accessibilityIdentifier("manual_discard_button")
-        .accessibilityLabel(Text("手札を捨て札にする"))
-        .accessibilityHint(Text(viewModel.manualDiscardAccessibilityHint))
-    }
-
-    /// 手動ペナルティ発動ボタン
-    private var manualPenaltyButton: some View {
-        let isDisabled = !viewModel.isManualPenaltyButtonEnabled
-
-        return Button {
-            viewModel.requestManualPenalty()
-        } label: {
-            Image(systemName: "arrow.triangle.2.circlepath")
-                .font(.system(size: 20, weight: .semibold))
-                .foregroundColor(theme.menuIconForeground)
-                .frame(width: 44, height: 44)
-                .background(
-                    Circle()
-                        .fill(theme.menuIconBackground)
-                )
-                .overlay(
-                    Circle()
-                        .stroke(theme.menuIconBorder, lineWidth: 1)
-                )
-        }
-        .buttonStyle(.plain)
-        .opacity(isDisabled ? 0.45 : 1.0)
-        .disabled(isDisabled)
-        .accessibilityIdentifier("manual_penalty_button")
-        .accessibilityLabel(Text("ペナルティを払って手札スロットを引き直す"))
-        .accessibilityHint(Text(viewModel.manualPenaltyAccessibilityHint))
-    }
-
-    /// ポーズメニュー表示ボタン
-    private var pauseButton: some View {
-        Button {
-            viewModel.presentPauseMenu()
-        } label: {
-            Image(systemName: "pause.circle")
-                .font(.system(size: 22, weight: .semibold))
-                .foregroundColor(theme.menuIconForeground)
-                .frame(width: 44, height: 44)
-                .background(
-                    Circle()
-                        .fill(theme.menuIconBackground)
-                )
-                .overlay(
-                    Circle()
-                        .stroke(theme.menuIconBorder, lineWidth: 1)
-                )
-        }
-        .buttonStyle(.plain)
-        .accessibilityIdentifier("pause_menu_button")
-        .accessibilityLabel(Text("ポーズメニュー"))
-        .accessibilityHint(Text("プレイを一時停止して設定やリセットを確認します"))
-    }
-
-    /// 統計バッジ 1 枚分の共通レイアウト
-    private func statisticBadge(
-        title: String,
-        value: String,
-        accessibilityLabel: String,
-        accessibilityValue: String
-    ) -> some View {
-        VStack(alignment: .leading, spacing: 2) {
-            Text(title)
-                .font(.caption2)
-                .fontWeight(.medium)
-                .foregroundColor(theme.statisticTitleText)
-
-            Text(value)
-                .font(.headline)
-                .foregroundColor(theme.statisticValueText)
-        }
-        .accessibilityElement(children: .ignore)
-        .accessibilityLabel(accessibilityLabel)
-        .accessibilityValue(accessibilityValue)
-    }
-
-    /// 経過時間の表示用フォーマット
-    private func formattedElapsedTime(_ seconds: Int) -> String {
-        let minutes = seconds / 60
-        let remainingSeconds = seconds % 60
-        return String(format: "%d:%02d", minutes, remainingSeconds)
-    }
-
-    /// 経過時間のアクセシビリティ説明
-    private func accessibilityElapsedTimeDescription(_ seconds: Int) -> String {
-        let hours = seconds / 3600
-        let minutes = (seconds % 3600) / 60
-        let remainingSeconds = seconds % 60
-
-        if hours > 0 {
-            return "\(hours)時間\(minutes)分\(remainingSeconds)秒"
-        } else if minutes > 0 {
-            return "\(minutes)分\(remainingSeconds)秒"
-        } else {
-            return "\(remainingSeconds)秒"
-        }
-    }
-
-    /// スコアのアクセシビリティ説明
-    private func accessibilityScoreDescription(_ score: Int) -> String {
-        "\(score)ポイント"
-    }
-}
-
-// MARK: - カード移動演出用オーバーレイ
-/// SpriteView と手札スロットの間を移動するカードのアニメーションを担当する補助ビュー
-/// - Note: GameView 本体のメソッドから切り出し、責務を明確にして可読性を高める。
-private struct GameCardAnimationOverlay: View {
-    /// 手札側で計測したアンカー情報の辞書
-    let anchors: [UUID: Anchor<CGRect>]
-    /// SpriteKit との橋渡しを担う ViewModel
-    @ObservedObject var boardBridge: GameBoardBridgeViewModel
-    /// フォールバック用に保持しておく現在地
-    /// - Note: アニメーションターゲットが未設定の場合に利用する
-    let fallbackCurrentPosition: GridPoint?
-    /// MatchedGeometryEffect の名前空間
-    let cardAnimationNamespace: Namespace.ID
-
-    var body: some View {
-        GeometryReader { proxy in
-            ZStack {
-                Color.clear
-                overlayContent(using: proxy)
-            }
-            .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .topLeading)
-        }
-    }
-
-    /// GeometryProxy を用いてカードの移動演出を構築する
-    /// - Parameter proxy: 手札・盤面それぞれの CGRect を解決するための GeometryProxy
-    @ViewBuilder
-    private func overlayContent(using proxy: GeometryProxy) -> some View {
-        if let animatingCard = boardBridge.animatingCard,
-           let sourceAnchor = anchors[animatingCard.id],
-           let boardAnchor = boardBridge.boardAnchor,
-           let targetGridPoint = boardBridge.animationTargetGridPoint ?? fallbackCurrentPosition,
-           boardBridge.animationState != .idle || boardBridge.hiddenCardIDs.contains(animatingCard.id) {
-            let cardFrame = proxy[sourceAnchor]
-            let boardFrame = proxy[boardAnchor]
-            let startCenter = CGPoint(x: cardFrame.midX, y: cardFrame.midY)
-            let boardDestination = Self.boardCoordinate(
-                for: targetGridPoint,
-                boardSize: boardBridge.boardSize,
-                in: boardFrame
-            )
-
-            MoveCardIllustrationView(card: animatingCard.move)
-                .matchedGeometryEffect(id: animatingCard.id, in: cardAnimationNamespace)
-                .frame(width: cardFrame.width, height: cardFrame.height)
-                .position(boardBridge.animationState == .movingToBoard ? boardDestination : startCenter)
-                .scaleEffect(boardBridge.animationState == .movingToBoard ? 0.55 : 1.0)
-                .opacity(boardBridge.animationState == .movingToBoard ? 0.0 : 1.0)
-                .allowsHitTesting(false)
-        }
-    }
-
-    /// 盤面座標を SwiftUI 座標系へ変換し、MatchedGeometryEffect の移動先を算出する
-    /// - Parameters:
-    ///   - gridPoint: 盤面上のマス座標（原点は左下）
-    ///   - boardSize: 現在の盤面一辺サイズ
-    ///   - frame: SwiftUI における盤面矩形
-    /// - Returns: SwiftUI 座標系での中心位置
-    private static func boardCoordinate(for gridPoint: GridPoint, boardSize: Int, in frame: CGRect) -> CGPoint {
-        // 盤面サイズが 0 以下になることは想定していないが、安全のため 1 以上へ補正する
-        let safeBoardSize = max(1, boardSize)
-        let tileLength = frame.width / CGFloat(safeBoardSize)
-        let centerX = frame.minX + tileLength * (CGFloat(gridPoint.x) + 0.5)
-        // SwiftUI 座標では上方向がマイナス値となるため、盤面上端からの距離で算出する
-        let centerY = frame.maxY - tileLength * (CGFloat(gridPoint.y) + 0.5)
-        return CGPoint(x: centerX, y: centerY)
-    }
-}
-
 // MARK: - コントロールバーの操作要素
 private extension GameView {
     /// メニュー操作を実際に実行する共通処理
     /// - Parameter action: ユーザーが選択した操作種別
     private func performMenuAction(_ action: GameMenuAction) {
         viewModel.performMenuAction(action)
-    }
-}
-
-// MARK: - レギュラー幅向けのメニュー確認シート
-/// iPad で確認文をゆったり表示するためのシートビュー
-// MARK: - ファイル内限定で利用する確認用シート
-/// `GameMenuAction` が fileprivate な型のため、このシートも fileprivate としておく
-fileprivate struct GameMenuActionConfirmationSheet: View {
-    /// 共通配色を参照して背景色などを統一する
-    private var theme = AppTheme()
-    /// 現在確認中のアクション
-    let action: GameMenuAction
-    /// 確定時に GameView 側で処理を実行するクロージャ
-    let onConfirm: (GameMenuAction) -> Void
-    /// キャンセル時に状態をリセットするクロージャ
-    let onCancel: () -> Void
-
-    /// 明示的なイニシャライザを用意し、`GameMenuAction` が `private` スコープでも呼び出せるようにする
-    /// - Parameters:
-    ///   - action: 確認対象のメニューアクション
-    ///   - onConfirm: 決定時に呼び出すクロージャ
-    ///   - onCancel: キャンセル時に呼び出すクロージャ
-    fileprivate init(
-        action: GameMenuAction,
-        onConfirm: @escaping (GameMenuAction) -> Void,
-        onCancel: @escaping () -> Void
-    ) {
-        // メンバーごとに代入し、従来通りの挙動を維持する
-        self.action = action
-        self.onConfirm = onConfirm
-        self.onCancel = onCancel
-    }
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 32) {
-            // MARK: - 見出しと詳細説明
-            VStack(alignment: .leading, spacing: 16) {
-                Text("操作の確認")
-                    .font(.title2)
-                    .fontWeight(.semibold)
-                    .frame(maxWidth: .infinity, alignment: .leading)
-                Text(action.confirmationMessage)
-                    .font(.body)
-                    .multilineTextAlignment(.leading)
-                    // iPad で視線移動が極端にならないように最大幅を確保しつつ左寄せで表示する
-                    .frame(maxWidth: .infinity, alignment: .leading)
-                    .foregroundStyle(theme.textPrimary)
-            }
-
-            Spacer(minLength: 0)
-
-            // MARK: - アクションボタン群
-            HStack(spacing: 16) {
-                Button("キャンセル", role: .cancel) {
-                    // ユーザーが操作を取り消した場合はシートを閉じる
-                    onCancel()
-                }
-                .buttonStyle(.bordered)
-
-                Button(action.confirmationButtonTitle, role: action.buttonRole) {
-                    // GameView 側で用意した実処理を実行する
-                    onConfirm(action)
-                }
-                .buttonStyle(.borderedProminent)
-            }
-            // ボタン行を右寄せにして重要ボタンへ視線を誘導する
-            .frame(maxWidth: .infinity, alignment: .trailing)
-        }
-        // iPad では余白を広めに確保し、モーダル全体が中央にまとまるようにする
-        .padding(32)
-        .frame(maxWidth: 520, alignment: .leading)
-        .background(theme.backgroundElevated)
-    }
-}
-
-/// ポーズメニュー本体。プレイ中によく調整する項目をリスト形式でまとめる
-private struct PauseMenuView: View {
-    /// カラーテーマを共有し、背景色やボタン色を統一する
-    private var theme = AppTheme()
-    /// プレイ再開ボタン押下時の処理
-    let onResume: () -> Void
-    /// リセット確定時の処理
-    let onConfirmReset: () -> Void
-    /// タイトルへ戻る確定時の処理
-    let onConfirmReturnToTitle: () -> Void
-
-    /// GameView 側から利用できるようアクセスレベルを内部公開にしたカスタムイニシャライザ
-    /// - Parameters:
-    ///   - onResume: ポーズ解除時に実行するクロージャ
-    ///   - onConfirmReset: ゲームリセット確定時に実行するクロージャ
-    ///   - onConfirmReturnToTitle: タイトル復帰確定時に実行するクロージャ
-    /// - Note: `private struct` では自動生成イニシャライザが private になるため、ここで明示的に定義する
-    init(
-        onResume: @escaping () -> Void,
-        onConfirmReset: @escaping () -> Void,
-        onConfirmReturnToTitle: @escaping () -> Void
-    ) {
-        self.onResume = onResume
-        self.onConfirmReset = onConfirmReset
-        self.onConfirmReturnToTitle = onConfirmReturnToTitle
-    }
-
-    /// シートを閉じるための環境ディスミス
-    @Environment(\.dismiss) private var dismiss
-    /// テーマ設定の永続化キー
-    @AppStorage("preferred_color_scheme") private var preferredColorSchemeRawValue: String = ThemePreference.system.rawValue
-    /// ハプティクスのオン/オフ
-    @AppStorage("haptics_enabled") private var hapticsEnabled: Bool = true
-    /// ガイドモードのオン/オフ
-    @AppStorage("guide_mode_enabled") private var guideModeEnabled: Bool = true
-    /// 手札並び設定
-    @AppStorage(HandOrderingStrategy.storageKey) private var handOrderingRawValue: String = HandOrderingStrategy.insertionOrder.rawValue
-
-    /// 破壊的操作の確認用ステート
-    @State private var pendingAction: PauseConfirmationAction?
-
-    var body: some View {
-        NavigationStack {
-            List {
-                // MARK: - プレイ再開ボタン
-                Section {
-                    Button {
-                        // フルスクリーンカバーを閉じて直ちにプレイへ戻る
-                        onResume()
-                        dismiss()
-                    } label: {
-                        Label("プレイを再開", systemImage: "play.fill")
-                    }
-                    .accessibilityHint("ポーズを解除してゲームを続けます")
-                }
-
-                // MARK: - ゲーム設定セクション
-                Section {
-                    Picker(
-                        "テーマ",
-                        selection: Binding<ThemePreference>(
-                            get: { ThemePreference(rawValue: preferredColorSchemeRawValue) ?? .system },
-                            set: { preferredColorSchemeRawValue = $0.rawValue }
-                        )
-                    ) {
-                        ForEach(ThemePreference.allCases) { preference in
-                            Text(preference.displayName)
-                                .tag(preference)
-                        }
-                    }
-
-                    Toggle("ハプティクスを有効にする", isOn: $hapticsEnabled)
-                    Toggle("ガイドモード（移動候補をハイライト）", isOn: $guideModeEnabled)
-
-                    Picker(
-                        "手札の並び順",
-                        selection: Binding<HandOrderingStrategy>(
-                            get: { HandOrderingStrategy(rawValue: handOrderingRawValue) ?? .insertionOrder },
-                            set: { handOrderingRawValue = $0.rawValue }
-                        )
-                    ) {
-                        ForEach(HandOrderingStrategy.allCases, id: \.self) { strategy in
-                            Text(strategy.displayName)
-                                .tag(strategy)
-                        }
-                    }
-                } header: {
-                    Text("ゲーム設定")
-                } footer: {
-                    Text("テーマやハプティクス、ガイド表示を素早く切り替えられます。これらの項目はタイトル画面の設定からも調整できます。")
-                }
-
-                // MARK: - 操作セクション
-                Section {
-                    Button(role: .destructive) {
-                        pendingAction = .reset
-                    } label: {
-                        Label("ゲームをリセット", systemImage: "arrow.counterclockwise")
-                    }
-
-                    Button(role: .destructive) {
-                        pendingAction = .returnToTitle
-                    } label: {
-                        Label("タイトルへ戻る", systemImage: "house")
-                    }
-                } header: {
-                    Text("操作")
-                } footer: {
-                    Text("リセットやタイトル復帰は確認ダイアログを経由して実行します。")
-                }
-
-                // MARK: - 詳細設定についての案内
-                Section {
-                    Text("広告やプライバシー設定などの詳細はタイトル画面右上のギアアイコンから確認できます。")
-                        .foregroundStyle(.secondary)
-                        .padding(.vertical, 4)
-                } header: {
-                    Text("詳細設定")
-                }
-            }
-            .listStyle(.insetGrouped)
-            .navigationTitle("ポーズ")
-            .toolbar {
-                ToolbarItem(placement: .cancellationAction) {
-                    Button("閉じる") {
-                        onResume()
-                        dismiss()
-                    }
-                }
-            }
-            .background(theme.backgroundPrimary)
-        }
-        // 破壊的操作の確認ダイアログ
-        .confirmationDialog(
-            "操作の確認",
-            // item: バインディングが iOS 17 以降で非推奨となったため、
-            // Bool バインディング + presenting の組み合わせで明示的に制御する
-            isPresented: Binding(
-                get: {
-                    // pendingAction が存在する場合のみダイアログを表示
-                    pendingAction != nil
-                },
-                set: { isPresented in
-                    // ユーザー操作でダイアログが閉じられたら状態を初期化
-                    if !isPresented {
-                        pendingAction = nil
-                    }
-                }
-            ),
-            presenting: pendingAction
-        ) { action in
-            // 確認用の破壊的操作ボタン
-            Button(action.confirmationButtonTitle, role: .destructive) {
-                handleConfirmation(action)
-            }
-            // キャンセルボタンは常に閉じるだけで状態を破棄
-            Button("キャンセル", role: .cancel) {
-                pendingAction = nil
-            }
-        } message: { action in
-            Text(action.message)
-        }
-    }
-
-    /// 確認ダイアログで選ばれたアクションを実行する
-    /// - Parameter action: ユーザーが確定した操作種別
-    private func handleConfirmation(_ action: PauseConfirmationAction) {
-        switch action {
-        case .reset:
-            onConfirmReset()
-            dismiss()
-        case .returnToTitle:
-            onConfirmReturnToTitle()
-            dismiss()
-        }
-        pendingAction = nil
-    }
-
-    /// ポーズメニュー内で扱う確認対象の列挙体
-    private enum PauseConfirmationAction: Identifiable {
-        case reset
-        case returnToTitle
-
-        var id: Int {
-            switch self {
-            case .reset: return 0
-            case .returnToTitle: return 1
-            }
-        }
-
-        var confirmationButtonTitle: String {
-            switch self {
-            case .reset: return "リセットする"
-            case .returnToTitle: return "タイトルへ戻る"
-            }
-        }
-
-        var message: String {
-            switch self {
-            case .reset:
-                return "現在の進行状況を破棄して最初からやり直します。よろしいですか？"
-            case .returnToTitle:
-                return "ゲームを終了してタイトル画面へ戻ります。現在のプレイ内容は保存されません。"
-            }
-        }
-    }
-}
-
-// MARK: - 手詰まりペナルティ用のバナー表示
-/// ペナルティ発生をユーザーに伝えるトップバナーのビュー
-private struct PenaltyBannerView: View {
-    /// バナー配色を一元管理するテーマ
-    private var theme = AppTheme()
-    /// 今回加算されたペナルティ手数
-    let penaltyAmount: Int
-
-    /// GameView 側から利用しやすいように外部公開されたイニシャライザを明示的に定義
-    /// - Parameter penaltyAmount: 今回のペナルティで増加した手数
-    init(penaltyAmount: Int) {
-        // `theme` はデフォルト値で初期化されるため代入不要。必要な値だけを受け取り保持する
-        self.penaltyAmount = penaltyAmount
-    }
-
-    var body: some View {
-        HStack(alignment: .center, spacing: 12) {
-            // MARK: - 警告アイコン
-            Image(systemName: "exclamationmark.triangle.fill")
-                .font(.system(size: 20, weight: .semibold))
-                // テーマの反転色を利用し、背景とのコントラストを確保
-                .foregroundColor(theme.penaltyIconForeground)
-                .padding(8)
-                .background(
-                    Circle()
-                        // アイコン背景もテーマ側のアクセントに合わせる
-                        .fill(theme.penaltyIconBackground)
-                )
-                .accessibilityHidden(true)  // アイコンは視覚的アクセントのみ
-
-            // MARK: - メッセージ本文
-            VStack(alignment: .leading, spacing: 2) {
-                Text(primaryMessage)
-                    .font(.system(size: 14, weight: .semibold, design: .rounded))
-                    // メインテキストはテーマに合わせた明度で表示
-                    .foregroundColor(theme.penaltyTextPrimary)
-                Text(secondaryMessage)
-                    .font(.system(size: 12, weight: .medium, design: .rounded))
-                    .foregroundColor(theme.penaltyTextSecondary)
-            }
-
-            Spacer(minLength: 0)
-        }
-        .padding(.vertical, 14)
-        .padding(.horizontal, 18)
-        .background(
-            RoundedRectangle(cornerRadius: 14, style: .continuous)
-                // バナーの背景はテーマ設定で透明感を調整
-                .fill(theme.penaltyBannerBackground)
-                .overlay(
-                    RoundedRectangle(cornerRadius: 14, style: .continuous)
-                        .stroke(theme.penaltyBannerBorder, lineWidth: 1)
-                )
-        )
-        .shadow(color: theme.penaltyBannerShadow, radius: 18, x: 0, y: 12)
-        .accessibilityIdentifier("penalty_banner_content")
-        .accessibilityElement(children: .combine)
-        .accessibilityLabel(accessibilityText)
-    }
-
-    /// ペナルティ内容に応じたメインメッセージ
-    private var primaryMessage: String {
-        if penaltyAmount > 0 {
-            return "手詰まり → 手札スロットを引き直し (+\(penaltyAmount))"
-        } else {
-            return "手札スロットを引き直しました (ペナルティなし)"
-        }
-    }
-
-    /// ペナルティ内容に応じた補足メッセージ
-    private var secondaryMessage: String {
-        if penaltyAmount > 0 {
-            return "使えるカードが無かったため、手数が \(penaltyAmount) 増加しました"
-        } else {
-            return "使えるカードが無かったため、手数の増加はありません"
-        }
-    }
-
-    /// アクセシビリティ用の案内文
-    private var accessibilityText: String {
-        "\(primaryMessage)。\(secondaryMessage)"
-    }
-}
-
-// MARK: - カード演出用の状態と PreferenceKey
-/// 手札・NEXT に配置されたカードのアンカーを UUID 単位で収集する PreferenceKey
-/// - Note: `GameHandSectionView` からも参照するためファイル外からアクセス可能にしている。
-struct CardPositionPreferenceKey: PreferenceKey {
-    static var defaultValue: [UUID: Anchor<CGRect>] = [:]
-
-    static func reduce(value: inout [UUID: Anchor<CGRect>], nextValue: () -> [UUID: Anchor<CGRect>]) {
-        value.merge(nextValue()) { _, new in new }
-    }
-}
-
-/// SpriteView（盤面）のアンカーを保持する PreferenceKey
-private struct BoardAnchorPreferenceKey: PreferenceKey {
-    static var defaultValue: Anchor<CGRect>? = nil
-
-    static func reduce(value: inout Anchor<CGRect>?, nextValue: () -> Anchor<CGRect>?) {
-        value = nextValue() ?? value
-    }
-}
-
-// MARK: - レイアウト定数と PreferenceKey
-/// 統計バッジ領域の高さを親ビューへ伝搬するための PreferenceKey
-private struct StatisticsHeightPreferenceKey: PreferenceKey {
-    static var defaultValue: CGFloat = 0
-
-    static func reduce(value: inout CGFloat, nextValue: () -> CGFloat) {
-        // 毎フレームで統計領域の最新高さへ更新し、ジオメトリ変化に即応できるよう最大値ではなく直近の値を採用する
-        value = nextValue()
-    }
-}
-
-/// 手札セクションの高さを親ビューへ伝搬するための PreferenceKey
-/// - Note: 手札専用ビューでも利用するため internal 扱いとする。
-struct HandSectionHeightPreferenceKey: PreferenceKey {
-    static var defaultValue: CGFloat = 0
-
-    static func reduce(value: inout CGFloat, nextValue: () -> CGFloat) {
-        // 手札レイアウトのジオメトリ変化へ滑らかに追従するため、最大値の保持ではなく常に最新高さへ更新する
-        value = nextValue()
-    }
-}
-
-/// 任意の PreferenceKey へ高さを伝搬するゼロサイズのオーバーレイ
-/// - Note: GeometryReader を直接レイアウトへ配置すると親ビューいっぱいまで広がり、想定外の値になるため
-///         あえて Color.clear を 0 サイズへ縮めて高さだけを測定する。
-///         `GameHandSectionView` からも再利用するため internal 扱いとする。
-struct HeightPreferenceReporter<Key: PreferenceKey>: View where Key.Value == CGFloat {
-    var body: some View {
-        GeometryReader { proxy in
-            Color.clear
-                .frame(width: 0, height: 0)
-                .preference(key: Key.self, value: proxy.size.height)
-        }
-        .allowsHitTesting(false)  // あくまでレイアウト取得用のダミービューなので操作対象から除外する
     }
 }
